feat(actions): ✨ Add new actions for debugging, folder management, and file operations
