+++ Apps.Box/Actions/FolderActions.cs
@@ -1,0 +1,70 @@
+﻿using Apps.Box.Dtos;
+using Apps.Box.Models.Requests;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Actions;
+using Blackbird.Applications.Sdk.Common.Invocation;
+using Blackbird.Applications.SDK.Extensions.FileManagement.Interfaces;
+using Box.V2.Models;
+
+namespace Apps.Box.Actions;
+
+[ActionList("Folders")]
+public class FolderActions(InvocationContext invocationContext, IFileManagementClient fileManagementClient)
+    : BoxInvocable(invocationContext)
+{
+    [Action("Create folder", Description = "Create folder")]
+    public async Task<string> CreateFolder([ActionParameter] CreateFolderRequest input)
+    {
+        var items = await ExecuteWithErrorHandlingAsync(async () => await Client.FoldersManager.GetFolderItemsAsync(
+            input.ParentFolderId, 300, 0, sort: BoxSortBy.Name.ToString(),
+            direction: BoxSortDirection.DESC, fields: new[] { "id", "type", "name" }));
+
+        var folders = items.Entries.Where(i => i.Type == "folder").ToList();
+        if (folders.Any(x => x.Name == input.FolderName))
+        {
+            return folders.FirstOrDefault(x => x.Name == input.FolderName).Id;
+        }
+
+        var folderRequest = new BoxFolderRequest
+        {
+            Name = input.FolderName,
+            Parent = new BoxRequestEntity
+            {
+                Id = input.ParentFolderId
+            }
+        };
+        var folder =
+            await ExecuteWithErrorHandlingAsync(async () => await Client.FoldersManager.CreateAsync(folderRequest));
+        return folder.Id;
+    }
+
+    [Action("Delete folder", Description = "Delete folder")]
+    public async Task DeleteFolder([ActionParameter] DeleteDirectoryRequest input)
+    {
+        await ExecuteWithErrorHandlingAsync(async () => await Client.FoldersManager.DeleteAsync(input.FolderId));
+    }
+
+    [Action("Add collaborator to folder", Description = "Add collaborator to folder")]
+    public async Task<CollaborationDto> AddCollaboratorToFolder([ActionParameter] AddFolderCollaboratorRequest input)
+    {
+        var addCollaboratorRequest = new BoxCollaborationRequest
+        {
+            AccessibleBy = new BoxCollaborationUserRequest
+            {
+                Login = input.CollaboratorEmail,
+                Type = BoxType.user
+            },
+            Item = new BoxRequestEntity
+            {
+                Id = input.FolderId,
+                Type = BoxType.folder
+            },
+            Role = input.Role
+        };
+        
+        var collaboration = await ExecuteWithErrorHandlingAsync(async () =>
+            await Client.CollaborationsManager.AddCollaborationAsync(addCollaboratorRequest,
+                notify: input.NotifyCollaborator));
+        return new CollaborationDto(collaboration);
+    }
+}
