+++ Apps.Box/Auth/OAuth2/OAuth2AuthorizeService.cs
@@ -5,12 +5,9 @@ using Blackbird.Applications.Sdk.Utils.Extensions.String;
 
 namespace Apps.Box.Auth.OAuth2;
 
-public class OAuth2AuthorizeService : BaseInvocable, IOAuth2AuthorizeService
+public class OAuth2AuthorizeService(InvocationContext invocationContext)
+    : BaseInvocable(invocationContext), IOAuth2AuthorizeService
 {
-    public OAuth2AuthorizeService(InvocationContext invocationContext) : base(invocationContext)
-    {
-    }
-
     public string GetAuthorizationUrl(Dictionary<string, string> values)
     {
         string bridgeOauthUrl = $"{InvocationContext.UriInfo.BridgeServiceUrl.ToString().TrimEnd('/')}/oauth";
@@ -25,6 +22,7 @@ public class OAuth2AuthorizeService : BaseInvocable, IOAuth2AuthorizeService
             { "authorization_url", oauthUrl},
             { "actual_redirect_uri", InvocationContext.UriInfo.AuthorizationCodeRedirectUri.ToString() },
         };
+        
         return bridgeOauthUrl.WithQuery(parameters);
     }
 }
