+++ Apps.Box/Events/Polling/Models/ParentFolderInput.cs
@@ -1,0 +1,11 @@
+﻿using Apps.Box.DataSourceHandlers;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Dynamic;
+
+namespace Apps.Box.Events.Polling.Models;
+
+public class ParentFolderInput
+{
+    [Display("Parent folder ID"), DataSource(typeof(FolderDataSourceHandler))]
+    public string? FolderId { get; set; }
+}
