+++ Apps.Box/Models/Requests/DownloadFileRequest.cs
@@ -1,12 +1,12 @@
 ﻿using Apps.Box.DataSourceHandlers;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.FileStorage;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Dynamic;
 
 namespace Apps.Box.Models.Requests;
 
-public class DownloadFileRequest
+public class DownloadFileRequest : IDownloadFileInput
 {
-    [DataSource(typeof(FileDataSourceHandler))]
-    [Display("File ID")]
+    [Display("File ID"), DataSource(typeof(FileDataSourceHandler))]
     public string FileId { get; set; }
 }
