+++ Apps.Box/Models/Requests/UploadFileRequest.cs
@@ -1,15 +1,15 @@
 ﻿using Apps.Box.DataSourceHandlers;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.FileStorage;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Dynamic;
 using Blackbird.Applications.Sdk.Common.Files;
 
 namespace Apps.Box.Models.Requests;
 
-public class UploadFileRequest
+public class UploadFileRequest : IUploadFileInput
 {
     public FileReference File { get; set; }
 
-    [DataSource(typeof(FolderDataSourceHandler))]
-    [Display("Parent folder")]
+    [Display("Parent folder"), DataSource(typeof(FolderDataSourceHandler))]
     public string ParentFolderId { get; set; }
 }
