+++ Apps.Box/Models/Responses/DownloadFileResponse.cs
@@ -1,8 +1,9 @@
-﻿using Blackbird.Applications.Sdk.Common.Files;
+﻿using Blackbird.Applications.SDK.Blueprints.Interfaces.FileStorage;
+using Blackbird.Applications.Sdk.Common.Files;
 
 namespace Apps.Box.Models.Responses;
 
-public class DownloadFileResponse
+public class DownloadFileResponse : IDownloadFileOutput
 {
     public FileReference File { get; set; }
 }
